Add threshold for floating-point SOFT_RELU activation
Added missing threshold for calculating SOFT_RELU when SVE and CL implementations are used. As a result removed from the testing bounds for input values that were set to be in the interval [-40, 40].

Resolves: COMPMID-5658
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/8536
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>

+++ tests/validation/reference/ActivationLayer.h
@@ -64,7 +64,7 @@ inline T activate_float(T x, T a, T b, ActivationLayerInfo::ActivationFunction a
             ret = (x > 0) ? x : a * x;
             break;
         case ActivationLayerInfo::ActivationFunction::SOFT_RELU:
-            ret = std::log(static_cast<T>(1) + std::exp(x));
+            ret = std::log(static_cast<T>(1) + std::exp(static_cast<double>(x)));
             break;
         case ActivationLayerInfo::ActivationFunction::ELU:
             ret = (x > 0) ? x : a * (std::exp(x) - static_cast<T>(1));
